♻️ F10 : BASKET > Add useMenu.js
