+++ src/components/pages/order/OrderPage.js
@@ -4,9 +4,8 @@ import Main from './main/Main';
 import styled from 'styled-components';
 import { theme } from '../../../theme';
 import OrderContext from '../../../context/OrderContext';
-import { fakeMenu } from '../../../fakeData/fakeMenu';
 import { EMPTY_PRODUCT } from '../../../enums/product';
-import { deepClone } from '../../../utils/array';
+import { useMenu } from '../../../hooks/useMenu';
 
 const OrderPage = () => {
   //state
@@ -15,44 +14,10 @@ const OrderPage = () => {
   const [isCollapsed, setIsCollapsed] = useState(false)
   const [isAddSelected, setIsAddSelected] = useState(true)
   const [isEditSelected, setIsEditSelected] = useState(false)
-  const [ menu, setMenu ] = useState(fakeMenu.MEDIUM)
   const [newProduct, setNewProduct] = useState(EMPTY_PRODUCT)
   const [productSelected, setProductSelected] = useState(EMPTY_PRODUCT)
   const titleEditRef = useRef()
-
-  //comportements (gestionnaire de state ou "state handlers")
-
-  const handleAdd = (newProduct) => {
-    //copy du tableau
-    const menuCopy = deepClone(menu)
-    //manip de la copie du tab
-    const menuUpdated = [ newProduct,...menuCopy]
-    //utilisation du setter
-    setMenu(menuUpdated)
-  }
-  const handleDelete =(idOfProductToDelete) => {
-    //1; copy du state
-    const menuCopy = deepClone(menu)
-    //2. manip de la copie du state
-    const menuUpdated = menuCopy.filter((product) => product.id !== idOfProductToDelete)
-    //3.Update du state
-    setMenu(menuUpdated)
-  }
-
-  const handleEdit = (productBeingEdited) => {
-    // 1. copie du state
-    const menuCopy = deepClone(menu)
-    // 2. manip de la copie du state
-    const indexOfProductToEdit = menu.findIndex((menuProduct) => menuProduct.id === productBeingEdited.id )
-    menuCopy[indexOfProductToEdit] = productBeingEdited
-    // 3. update du state
-    setMenu(menuCopy)
-  }
-  
-
-  const resetMenu = () => {
-    setMenu(fakeMenu.MEDIUM)
-  }
+  const { menu, handleAdd, handleDelete, handleEdit, resetMenu } = useMenu()
   
   const orderContextValue = {
     isModeAdmin,
@@ -66,7 +31,6 @@ const OrderPage = () => {
     currentTabSelected,
     setCurrentTabSelected,
     menu,
-    setMenu,
     handleAdd,
     handleDelete,
     handleEdit,
