+++ src/components/pages/order/main/Basket/Basket.js
@@ -18,5 +18,4 @@ const BasketStyled = styled.div`
   background: pink;
   display: flex;
   flex-direction: column;
-
 `
